feat: add LRU list and listNode but LRU not support evict now

+++ util_hash.go
@@ -4,6 +4,6 @@ import "github.com/dolthub/maphash"
 
 var hasher = maphash.NewHasher[string]()
 
-var xxHashString = func(key string) uintptr {
-	return uintptr(hasher.Hash(key))
+var xxHashString = func(key string) uint64 {
+	return hasher.Hash(key)
 }
